Major overhaul
verifiy_certificate now behaves as a outside checker to validate assumptions in coq. This way a check on the whole 2^30 generated data set was possible. Not all implementations are perfect yet, I stopped at junctions where it "worked". Later updates will simplify the library. Please note at current the generator will generate around 120gb in batched certificates, obviously this will not be required in the future. the minimized.db and other bigger artifacts used for the current assumptions will be uploaded to zenodo